requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pymongo==4.10.1
python-dotenv==1.0.1
pydantic[email]==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
beautifulsoup4==4.12.3
pandas==2.2.3
openpyxl==3.1.5
httpx==0.28.1